Second Mind - Advance - Script Version/pyproject.toml
[project]
name = "knowledge-system"
version = "0.4.0"
description = "LLM-assisted ingestion for Obsidian wiki (raw → processed wiki → insights)."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pypdf>=5.0.0",
    "pyyaml>=6.0.0",
]

[project.scripts]
kns = "scripts.cli:main"
kns-ingest = "scripts.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scripts"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"
